feat: extend storage by default

+++ install.sh
@@ -453,28 +453,29 @@ EOF
       uci set fstab.@global[0].check_fs='1'
       uci commit fstab
       service fstab boot
-      ;;
-    "Extend-Storage")
-      opkg install block-mount kmod-fs-ext4 e2fsprogs parted
-      parted -s /dev/sda -- mklabel gpt mkpart extroot 2048s -2048s
-      DEVICE="$(block info | sed -n -e '/MOUNT="\S*\/overlay"/s/:\s.*$//p')"
-      uci -q delete fstab.rwm
-      uci set fstab.rwm="mount"
-      uci set fstab.rwm.device="${DEVICE}"
-      uci set fstab.rwm.target="/rwm"
-      uci commit fstab
-      DEVICE="/dev/sda1"
-      mkfs.ext4 -L extroot ${DEVICE}
-      eval $(block info ${DEVICE} | grep -o -e 'UUID="\S*"')
-      eval $(block info | grep -o -e 'MOUNT="\S*/overlay"')
-      uci -q delete fstab.extroot
-      uci set fstab.extroot="mount"
-      uci set fstab.extroot.uuid="${UUID}"
-      uci set fstab.extroot.target="${MOUNT}"
-      uci commit fstab
-      mount ${DEVICE} /mnt
-      tar -C ${MOUNT} -cvf - . | tar -C /mnt -xf -
-      reboot
+      read -r -p "Do You Want To Extend Storage? (yes/no): " EXTEND_STORAGE
+      if [[ "$EXTEND_STORAGE" != "no" ]]; then
+        opkg install block-mount kmod-fs-ext4 e2fsprogs parted
+        parted -s /dev/sda -- mklabel gpt mkpart extroot 2048s -2048s
+        DEVICE="$(block info | sed -n -e '/MOUNT="\S*\/overlay"/s/:\s.*$//p')"
+        uci -q delete fstab.rwm
+        uci set fstab.rwm="mount"
+        uci set fstab.rwm.device="${DEVICE}"
+        uci set fstab.rwm.target="/rwm"
+        uci commit fstab
+        DEVICE="/dev/sda1"
+        mkfs.ext4 -L extroot ${DEVICE}
+        eval $(block info ${DEVICE} | grep -o -e 'UUID="\S*"')
+        eval $(block info | grep -o -e 'MOUNT="\S*/overlay"')
+        uci -q delete fstab.extroot
+        uci set fstab.extroot="mount"
+        uci set fstab.extroot.uuid="${UUID}"
+        uci set fstab.extroot.target="${MOUNT}"
+        uci commit fstab
+        mount ${DEVICE} /mnt
+        tar -C ${MOUNT} -cvf - . | tar -C /mnt -xf -
+        reboot
+      fi
       ;;
     "Swap")
       opkg install zram-swap
@@ -489,7 +490,7 @@ EOF
 menu() {
   PS3="Enter Your Option: "
   OPTIONS=(
-    "Setup" "Upgrade" "Recommended" "LoadBalancer" "QoS" "AdGuard" "Passwall" "WapPlus" "Hiddify" "USB-WAN" "USB-Storage" "Extend-Storage" "Swap" "Quit"
+    "Setup" "Upgrade" "Recommended" "LoadBalancer" "QoS" "AdGuard" "Passwall" "WapPlus" "Hiddify" "USB-WAN" "USB-Storage" "Swap" "Quit"
   )
   select CHOICE in "${OPTIONS[@]}"; do
     run_commands "$CHOICE"
